- use binary type for ID column for pgsql
