+++ src/Symfony/Component/Cache/Tests/Adapter/TagAwareAdapterAndPdoAdapterTest.php
@@ -1,0 +1,53 @@
+<?php
+
+/*
+ * This file is part of the Symfony package.
+ *
+ * (c) Fabien Potencier <fabien@symfony.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Symfony\Component\Cache\Tests\Adapter;
+
+use Doctrine\DBAL\DriverManager;
+use PHPUnit\Framework\SkippedTestSuiteError;
+use Psr\Cache\CacheItemPoolInterface;
+use Symfony\Component\Cache\Adapter\AbstractAdapter;
+use Symfony\Component\Cache\Adapter\DoctrineDbalAdapter;
+use Symfony\Component\Cache\Adapter\PdoAdapter;
+use Symfony\Component\Cache\Adapter\TagAwareAdapter;
+use Symfony\Component\Cache\Tests\Adapter\TagAwareAdapterTestCase;
+
+/**
+ * @group time-sensitive
+ */
+class TagAwareAdapterAndPdoAdapterTest extends TagAwareAdapterTestCase
+{
+    protected static $dbFile;
+
+    public static function setUpBeforeClass(): void
+    {
+        if (!\extension_loaded('pdo_sqlite')) {
+            throw new SkippedTestSuiteError('Extension pdo_sqlite required.');
+        }
+
+        self::$dbFile = tempnam(sys_get_temp_dir(), 'sf_sqlite_cache');
+    }
+
+    public static function tearDownAfterClass(): void
+    {
+        @unlink(self::$dbFile);
+    }
+
+    public function createCachePool(int $defaultLifetime = 0): CacheItemPoolInterface
+    {
+        return new TagAwareAdapter(new PdoAdapter('sqlite:'.self::$dbFile, '', $defaultLifetime));
+    }
+
+    protected function createCacheAdapter(): AbstractAdapter
+    {
+        return new PdoAdapter('sqlite:'.self::$dbFile);
+    }
+}
